Avoid forward declaration as by including DocsumMetrics in definition.
