+++ searchcore/src/vespa/searchcore/proton/summaryengine/summaryengine.h
@@ -6,10 +6,11 @@
 #include <vespa/searchcore/proton/common/handlermap.hpp>
 #include <vespa/searchlib/engine/docsumapi.h>
 #include <vespa/vespalib/util/threadstackexecutor.h>
+#include <vespa/metrics/valuemetric.h>
+#include <vespa/metrics/countmetric.h>
+#include <vespa/metrics/metricset.h>
 #include <mutex>
 
-namespace metrics { class MetricSet; }
-namespace proton::internal { class DocsumMetrics; }
 
 namespace proton {
 
@@ -21,11 +22,20 @@ private:
     using DocsumRequest = search::engine::DocsumRequest;
     using DocsumClient = search::engine::DocsumClient;
 
+    struct DocsumMetrics : metrics::MetricSet {
+        metrics::LongCountMetric     count;
+        metrics::LongCountMetric     docs;
+        metrics::DoubleAverageMetric latency;
+
+        DocsumMetrics();
+        ~DocsumMetrics();
+    };
+
     std::mutex                    _lock;
     bool                          _closed;
     HandlerMap<ISearchHandler>    _handlers;
     vespalib::ThreadStackExecutor _executor;
-    std::unique_ptr<internal::DocsumMetrics> _metrics;
+    std::unique_ptr<metrics::MetricSet> _metrics;
 
 public:
     SummaryEngine(const SummaryEngine &) = delete;
@@ -106,7 +116,7 @@ public:
      */
     DocsumReply::UP getDocsums(DocsumRequest::UP req) override;
 
-    metrics::MetricSet & getMetrics();
+    metrics::MetricSet & getMetrics() { return *_metrics; }
 };
 
 } // namespace proton
